feat(testing): Moved language actions to ProjectsPage
